+++ 2Week_Mission/src/main/java/com/ll/exam/ebooks/app/order/entity/OrderItem.java
@@ -64,4 +64,8 @@ public class OrderItem extends BaseEntity {
         this.isPaid = true;
         this.payDate = LocalDateTime.now();
     }
+
+    public void setRefundDone() {
+        this.refundPrice = payPrice;
+    }
 }
