+++ 2Week_Mission/src/main/java/com/ll/exam/ebooks/app/order/exception/ActorCanNotRefundException.java
@@ -1,0 +1,4 @@
+package com.ll.exam.ebooks.app.order.exception;
+
+public class ActorCanNotRefundException extends RuntimeException {
+}
