+++ 2Week_Mission/src/main/resources/templates/order/detail.html
@@ -96,7 +96,19 @@
                                 <form th:action="@{|/order/${order.id}/cancel|}" method="POST" hidden></form>
                             </div>
                         </div>
-                        <div th:if="${order.isPaid}" class="mt-2">
+                        <div th:if="${order.isPaid}" class="grid grid-cols-2 my-3 gap-2">
+                            <button class="btn btn-mb btn-primary" disabled="disabled">
+                                <i class="fa-regular fa-credit-card"></i>
+                                <span class="ml-1">결제완료</span>
+                            </button>
+                            <a href="javascript:;" onclick="if ( confirm('환불하시겠습니까?') ) $(this).next().submit();"
+                               class="btn btn-secondary btn-mb btn-outline">
+                                <i class="fa-solid fa-rotate-left"></i>
+                                <span class="ml-1">결제환불</span>
+                            </a>
+                            <form th:action="@{|/order/${order.id}/refund|}" method="POST" hidden></form>
+                        </div>
+                        <div th:if="${order.isPaid()}">
                             <div class="badge badge-lg badge-secondary">결제완료</div>
                             <div class="mt-3">
                                 결제 내역
@@ -117,7 +129,7 @@
         </div>
     </section>
 
-    <script>
+    <script th:inline="javascript">
         let $PaymentForm__useRestCash = $("#PaymentForm__useRestCash");
         const $PaymentForm__useRestCashAll = $("#PaymentForm__useRestCashAll");
         const $PaymentForm__pgPayPriceDisplay = $("#PaymentForm__pgPayPriceDisplay");
@@ -142,22 +154,18 @@
                 orderId: "order__" + orderId + "__" + (Math.random() + "").substring(2),
                 orderName,
                 customerName: buyerName,
-                successUrl: window.location.origin + "/order/" + orderId + "/success",
-                failUrl: window.location.origin + "/order/" + orderId + "/fail"
+                successUrl: window.location.origin + "/toss/" + orderId + "/success",
+                failUrl: window.location.origin + "/toss/" + orderId + "/fail"
             };
 
             tossPayments.requestPayment(method, paymentData);
         }
     </script>
 
-    <script>
+    <script th:inline="javascript">
         // 예치금 입력폼 최솟값(0), 최댓값(보유 예치금 및 결제 금액) 자동 설정
         function PaymentForm__useRestCashFix() {
             let useRestCash = parseInt($PaymentForm__useRestCash.val());
-
-            if (isNaN(useRestCash)) {
-                useRestCash = 0;
-            }
 
             const maxUseRestCash = parseInt($PaymentForm__useRestCash.attr('max'));
